\name{bayeslr-package}
\alias{bayeslr-package}
\alias{bayeslr}
\docType{package}
\title{
\packageTitle{bayeslr}
}
\description{
\packageDescription{bayeslr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bayeslr}
\packageIndices{bayeslr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bayeslr}

Maintainer: \packageMaintainer{bayeslr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
